Refactor

+++ Tests/Source/LinearAllocatorTest.cpp
@@ -42,26 +42,26 @@ Object* CheckNewArrayRaw(LinearAllocator<>& allocator, size_t objectCount, Args&
 
 TEST_F(LinearAllocatorTest, Initialize) { EXPECT_EQ(linearAllocator.GetUsedSize(), 0); }
 
-TEST_F(LinearAllocatorTest, RawNewSingleObject) { CheckNewRaw<TestObject>(linearAllocator, 1, 2.1f, 'a', false, 10.6f); }
+TEST_F(LinearAllocatorTest, RawNewSingleObject) { CheckNewRaw<TestObject>(linearAllocator, 1, 2.1F, 'a', false, 10.6f); }
 
 TEST_F(LinearAllocatorTest, RawNewMultipleObjects)
 {
     for (size_t i = 0; i < 10; i++)
     {
-        CheckNewRaw<TestObject>(linearAllocator, i, i + 1.5f, 'a' + i, i % 2, i + 2.5f);
+        CheckNewRaw<TestObject>(linearAllocator, i, static_cast<float>(i) + 1.5F, 'a' + i, i % 2, static_cast<float>(i) + 2.5F);
     }
     for (size_t i = 0; i < 10; i++)
     {
-        CheckNewRaw<TestObject2>(linearAllocator, i, i + 1.5, i + 2.5, i % 2, Pair{1, 2.5f});
+        CheckNewRaw<TestObject2>(linearAllocator, i, static_cast<float>(i) + 1.5, static_cast<float>(i) + 2.5, i % 2, Pair{1, 2.5F});
     }
 }
 
 TEST_F(LinearAllocatorTest, RawNewMixed)
 {
-    TestObject* arr1    = CheckNewArrayRaw<TestObject>(linearAllocator, 10, 1, 2.1f, 'a', false, 10.6f);
-    TestObject* object1 = CheckNewRaw<TestObject>(linearAllocator, 1, 2.1f, 'a', false, 10.6f);
-    TestObject* object2 = CheckNewRaw<TestObject>(linearAllocator, 1, 2.1f, 'a', false, 10.6f);
-    TestObject* arr2    = CheckNewArrayRaw<TestObject>(linearAllocator, 10, 1, 2.1f, 'a', false, 10.6f);
+    TestObject* arr1    = CheckNewArrayRaw<TestObject>(linearAllocator, 10, 1, 2.1F, 'a', false, 10.6f);
+    TestObject* object1 = CheckNewRaw<TestObject>(linearAllocator, 1, 2.1F, 'a', false, 10.6f);
+    TestObject* object2 = CheckNewRaw<TestObject>(linearAllocator, 1, 2.1F, 'a', false, 10.6f);
+    TestObject* arr2    = CheckNewArrayRaw<TestObject>(linearAllocator, 10, 1, 2.1F, 'a', false, 10.6f);
 }
 
 TEST_F(LinearAllocatorTest, Release)
@@ -69,14 +69,16 @@ TEST_F(LinearAllocatorTest, Release)
     LinearAllocator<> linearAllocator2 = LinearAllocator<>(10 * sizeof(TestObject));
     for (size_t i = 0; i < 10; i++)
     {
-        TestObject* object = CheckNewRaw<TestObject>(linearAllocator2, i, i + 1.5f, 'a' + i, i % 2, i + 2.5f);
+        TestObject* object =
+            CheckNewRaw<TestObject>(linearAllocator2, i, static_cast<float>(i) + 1.5F, 'a' + i, i % 2, static_cast<float>(i) + 2.5F);
     }
 
     linearAllocator2.Release();
 
     for (size_t i = 0; i < 10; i++)
     {
-        TestObject* object = CheckNewRaw<TestObject>(linearAllocator2, i, i + 1.5f, 'a' + i, i % 2, i + 2.5f);
+        TestObject* object =
+            CheckNewRaw<TestObject>(linearAllocator2, i, static_cast<float>(i) + 1.5F, 'a' + i, i % 2, static_cast<float>(i) + 2.5F);
     }
 }
 
@@ -85,7 +87,8 @@ TEST_F(LinearAllocatorTest, GetUsedSizeNew)
     const int numObjects = 10;
     for (size_t i = 0; i < numObjects; i++)
     {
-        TestObject* object = CheckNewRaw<TestObject>(linearAllocator, i, i + 1.5f, 'a' + i, i % 2, i + 2.5f);
+        TestObject* object =
+            CheckNewRaw<TestObject>(linearAllocator, i, static_cast<float>(i) + 1.5F, 'a' + i, i % 2, static_cast<float>(i) + 2.5F);
     }
 
     EXPECT_EQ(linearAllocator.GetUsedSize(), numObjects * (sizeof(TestObject)));
@@ -95,7 +98,7 @@ TEST_F(LinearAllocatorTest, GetUsedSizeNewArray)
 {
     const int numObjects = 10;
 
-    TestObject* arr = CheckNewArrayRaw<TestObject>(linearAllocator, numObjects, 1, 2.1f, 'a', false, 10.6f);
+    TestObject* arr = CheckNewArrayRaw<TestObject>(linearAllocator, numObjects, 1, 2.1F, 'a', false, 10.6f);
 
     EXPECT_EQ(linearAllocator.GetUsedSize(), std::max(alignof(TestObject), numObjects * sizeof(TestObject)));
 }
@@ -107,8 +110,8 @@ void ThreadFunction(LinearAllocator<policy>& linearAllocator)
 
     for (size_t i = 0; i < 10000; i++)
     {
-        TestObject* testObject = linearAllocator.template NewRaw<TestObject>(1, 1.5f, 'a', false, 2.5f);
-        EXPECT_EQ(*testObject, TestObject(1, 1.5F, 'a', false, 2.5f));
+        TestObject* testObject = linearAllocator.template NewRaw<TestObject>(1, 1.5F, 'a', false, 2.5F);
+        EXPECT_EQ(*testObject, TestObject(1, 1.5F, 'a', false, 2.5F));
 
         objects.push_back(testObject);
     }
@@ -142,8 +145,8 @@ TEST_F(LinearAllocatorTest, Resizable)
     const int numObjects = 10;
     for (size_t i = 0; i < numObjects; i++)
     {
-        TestObject* testObject = linearAllocator2.NewRaw<TestObject>(1, 1.5f, 'a', false, 2.5f);
-        EXPECT_EQ(*testObject, TestObject(1, 1.5f, 'a', false, 2.5f));
+        TestObject* testObject = linearAllocator2.NewRaw<TestObject>(1, 1.5F, 'a', false, 2.5F);
+        EXPECT_EQ(*testObject, TestObject(1, 1.5F, 'a', false, 2.5F));
     }
 
     EXPECT_EQ(linearAllocator2.GetUsedSize(), (sizeof(TestObject) * 10));
@@ -153,11 +156,11 @@ TEST_F(LinearAllocatorTest, Templated)
 {
     LinearAllocatorTemplated<TestObject> linearAllocatorTemplated{10_KB};
 
-    TestObject* testObject = linearAllocatorTemplated.NewRaw(1, 1.5f, 'a', false, 2.5f);
-    EXPECT_EQ(*testObject, TestObject(1, 1.5f, 'a', false, 2.5f));
+    TestObject* testObject = linearAllocatorTemplated.NewRaw(1, 1.5F, 'a', false, 2.5F);
+    EXPECT_EQ(*testObject, TestObject(1, 1.5F, 'a', false, 2.5F));
 
-    TestObject* testObject3 = linearAllocatorTemplated.NewArrayRaw(10, 1, 1.5f, 'a', false, 2.5f);
-    EXPECT_EQ(*testObject, TestObject(1, 1.5f, 'a', false, 2.5f));
+    TestObject* testObject3 = linearAllocatorTemplated.NewArrayRaw(10, 1, 1.5F, 'a', false, 2.5F);
+    EXPECT_EQ(*testObject, TestObject(1, 1.5F, 'a', false, 2.5F));
 
     linearAllocatorTemplated.Release();
 
@@ -186,7 +189,7 @@ TEST_F(LinearAllocatorDeathTest, NewOutOfMemory)
     LinearAllocator linearAllocator2{10};
 
     // TODO Write proper exit messages
-    ASSERT_DEATH({ TestObject* object = linearAllocator2.NewRaw<TestObject>(1, 2.1f, 'a', false, 10.6f); }, ".*");
+    ASSERT_DEATH({ TestObject* object = linearAllocator2.NewRaw<TestObject>(1, 2.1F, 'a', false, 10.6f); }, ".*");
 }
 
 #endif
